module6-task1 fix
