+++ src/store/reducers/current-city.js
@@ -1,0 +1,14 @@
+import {ActionType} from '../actions';
+import {Cities} from '../../constants';
+
+const initialState = Cities.PARIS;
+
+export const currentCityName = (state = initialState, action) => {
+
+  switch (action.type) {
+    case ActionType.SET_CITY:
+      return action.payload;
+
+    default: return state;
+  }
+};
