+++ src/store/reducers/favorites.js
@@ -1,0 +1,8 @@
+// import {ActionType} from '../actions';
+import {Favorites} from '../../mocks/favorites';
+
+const initialState = Favorites;
+
+export const favorites = (state = initialState, _action) => {
+  return state;
+};
