+++ src/store/reducers/reducer.js
@@ -1,51 +1,15 @@
-import {ActionType} from '../actions';
-import {Favorites} from '../../mocks/favorites';
-import {Cities, SortOrders, AuthorizationStatus} from '../../constants';
+import {combineReducers} from 'redux';
 
-const initialState = {
-  currentCityName: Cities.PARIS,
-  sortOrder: SortOrders.POPULAR,
-  offers: {
-    isLoaded: false,
-    entities: [],
-  },
-  authorizationStatus: AuthorizationStatus.NO_AUTH,
-  favorites: Favorites
-};
+import {offers} from './offers';
+import {authorizationStatus} from './auth-status';
+import {sortOrder} from './sort-order';
+import {currentCityName} from './current-city';
+import {favorites} from './favorites';
 
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-
-    case ActionType.SET_CITY:
-      return {
-        ...state,
-        currentCityName: action.payload
-      };
-
-    case ActionType.SET_SORT_ORDER:
-      return {
-        ...state,
-        sortOrder: action.payload
-      };
-
-    case ActionType.LOAD_OFFERS:
-      return {
-        ...state,
-        offers: {
-          ...state.offers,
-          isLoaded: true,
-          entities: action.payload,
-        }
-      };
-
-    case ActionType.REQUIRED_AUTHORIZATION:
-      return {
-        ...state,
-        authorizationStatus: action.payload
-      };
-
-    default: return state;
-  }
-};
-
-export {reducer};
+export const reducer = combineReducers({
+  currentCityName,
+  sortOrder,
+  offers,
+  authorizationStatus,
+  favorites,
+});
